pyproject.toml
[project]
name = "line"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.7",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "decorator==5.2.1",
    "executing==2.2.0",
    "filetype==1.2.0",
    "frozenlist==1.6.0",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.24.2",
    "google-auth==2.40.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.25",
    "langchain-core==0.3.63",
    "langchain-google-genai==2.1.5",
    "langchain-tavily==0.2.0",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.7",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.44",
    "load-dotenv==0.1.0",
    "matplotlib-inline==0.1.7",
    "multidict==6.4.4",
    "mypy==1.16.0",
    "mypy-extensions==1.1.0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "parso==0.8.4",
    "pathspec==0.12.1",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "traitlets==5.14.3",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]
